requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.1
fpdf==1.7.2
flask==2.3.2
psycopg2-binary==2.9.6